backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-google-genai
langchain-core
chromadb
pydantic
